Added automatic update
