+++ collect_app/src/main/java/org/odk/collect/android/listeners/FormDownloaderListener.java
@@ -1,31 +1,24 @@
 /*
- * Copyright (C) 2009 University of Washington
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * 
+ * Copyright 2018 Nafundi
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.odk.collect.android.listeners;
 
-import org.odk.collect.android.logic.FormDetails;
-
-import java.util.HashMap;
-
-
-/**
- * @author Carl Hartung (carlhartung@gmail.com)
- */
 public interface FormDownloaderListener {
-    void formsDownloadingComplete(HashMap<FormDetails, String> result);
 
-    void progressUpdate(String currentFile, int progress, int total);
+    void progressUpdate(String currentFile, String progress, String total);
 
-    void formsDownloadingCancelled();
+    boolean isTaskCanceled();
 }
