+++ collect_app/src/main/res/xml/form_management_preferences.xml
@@ -60,4 +60,8 @@
             android:key="periodic_form_updates_check"
             android:title="@string/periodic_form_updates_check_title" />
     </PreferenceCategory>
+    <CheckBoxPreference
+        android:key="automatic_update"
+        android:summary="@string/automatic_update_summary"
+        android:title="@string/automatic_update" />
 </PreferenceScreen>
