+++ collect_app/src/main/res/xml/user_settings_access_preferences.xml
@@ -29,6 +29,9 @@
             android:key="periodic_form_updates_check"
             android:title="@string/periodic_form_updates_check_title" />
         <CheckBoxPreference
+            android:key="automatic_update"
+            android:title="@string/automatic_update" />
+        <CheckBoxPreference
             android:key="default_to_finalized"
             android:title="@string/default_completed" />
         <CheckBoxPreference
